Command Bar clean up (#2108)
* refactor command-bar API
   ```diff
   - <CommandBarDialog isOpen={state.isOpen} onDismiss={toggleIsOpen}>
   + <Command.Dialog isOpen={state.isOpen} onDismiss={toggleIsOpen}>
   +   <Command.Header />
   +   <Command.Body />
   +   <Command.Footer />
   + </Command.Dialog>
   ```
* consolidate & delete `types` files
