--- src/components/command-bar/components/activator/types.ts
@@ -1,13 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-import { ReactElement } from 'react'
-
-interface CommandBarActivatorProps {
-	leadingIcon: ReactElement
-	visualLabel: string
-}
-
-export type { CommandBarActivatorProps }
