--- src/components/command-bar/components/dialog/types.ts
@@ -1,15 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-interface CommandBarDialogProps {
-	isOpen?: boolean
-	onDismiss?: () => void
-}
-
-interface CommandBarDialogFooterProps {
-	instructionsElementId: string
-}
-
-export type { CommandBarDialogFooterProps, CommandBarDialogProps }
