--- src/components/command-bar/components/list-item/types.ts
@@ -1,34 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-import { ReactElement, ReactNode } from 'react'
-import { BadgeProps } from 'components/badge'
-
-interface CommandBarListItemProps {
-	children: ReactNode
-}
-
-interface CommandBarListItemContentProps {
-	badges?: BadgeProps['text'][]
-	description?: string
-	icon?: ReactElement
-	title: string
-	trailingIcon?: ReactElement
-}
-
-interface CommandBarButtonListItemProps extends CommandBarListItemContentProps {
-	onClick: () => void
-}
-
-interface CommandBarLinkListItemProps extends CommandBarListItemContentProps {
-	url: string
-}
-
-export type {
-	CommandBarButtonListItemProps,
-	CommandBarLinkListItemProps,
-	CommandBarListItemContentProps,
-	CommandBarListItemProps,
-}
